requirements.txt
fastapi
groq
openai
python-dotenv
